requirements.txt
swig
Pillow
pygame
numpy 
matplotlib
scipy
opencv-python
gymnasium
gymnasium[box2d]
neat-python